+++ airport_generation.py
@@ -9,10 +9,9 @@ import string, random
 # continent codes https://en.wikipedia.org/wiki/ICAO_airport_code#/media/File:ICAO_FirstLetter.svg
 continentCodes = ["K", "C", "S", "U", "Z", "R", "Y", "V", "B", "O", "E", "L"]
 letters = list(string.ascii_uppercase)
-#airport = Airport("KLAX", [0,0], [], 'A', [330, 12])
 sizes = ["A", "B", "C", "D", "E", "F"]
 
-#* size guideline  
+#* size guideline (custom based on aircraft classes)
 #* farm, airstrip - A, regional - B, metropolitan - C, semi-international - D
 #* average international - E, major international - F
 
@@ -23,17 +22,19 @@ def generateCode():
         code += random.choice(letters)
     return code
 
+# generates airport size from classes "A" to "F"
 def generateSize(sizes):
     return random.choice(sizes)
 
+# generates runway count for airport based on size
 def runwayCount(size):
     mass = ord(size) - ord("A")
     if mass < 2:
         return 1
     return mass
 
-""" # runway parameters
-def newLength(airport):
+# varying runway length by airport size
+""" def newLength(airport):
     if ord(airport.size) < ord("B"):
         return random.randrange(500, 2000)
     elif ord(airport.size) < ord("C"):
@@ -42,22 +43,25 @@ def newLength(airport):
         diff = ord(airport.size) - ord("A")
         return random.randrange(6000 + diff * 1000, 7000 + diff * 1000) """
 
-def newRwyPos( hdg, length):
+# generates runway position based on radius vector from airport position
+def newRwyPos(hdg, length):
     normRwy = normalVector(hdgVector(hdg, length / 1500))
     pos = list(map(lambda x,y: x+y, [0,0], normRwy))
     return pos
 
+# generates runway with a heading tangent to the circle with the center airport
 def generateRunway(airport):
     hdg = random.randrange(0, 360)
     length = 11000
     pos = newRwyPos(hdg, length)
     for runway in airport.runways:
-        while distance(pos, runway.pos) < 4 and abs(hdg - runway.hdg) < 10:
+        while distance(pos, runway.pos) < 15 and abs(hdg - runway.hdg) < 20:
             hdg = random.randrange(0, 360)
             pos = newRwyPos(hdg, length)
     rwy = roundHalfUp(hdg / 10)
     return Runway(rwy, pos, hdg, length, airport)
 
+# sets wind for runways corresponding to wind map position
 def generateWind(runways):
     spd = random.randrange(2,15)
     avg = 0
@@ -65,6 +69,7 @@ def generateWind(runways):
             avg += runway.hdg
     return [int(avg / len(runways)), spd]
 
+# generates airport object with parameters generated above
 def generateAirport(pos):
     code = generateCode()
     size = generateSize(sizes)
@@ -75,8 +80,3 @@ def generateAirport(pos):
         #pprint(f"Runway: {vars(airport.runways[count])}")
     airport.wind = generateWind(runways)
     return airport
-
-
-"""         if runway.hdg > 180:
-            avg += ((runway.hdg + 180) % 360)
-        else: """