+++ commands.py
@@ -15,12 +15,15 @@ testFlights = [Flight('DAL123', 'B738', [0, 0], 234, 230, 9000, 0, 'KLAX', 'RKSI
 
 # typo recgnition
 
+# returns dictionary of letter counts
 def letterCounts(word):
     result = dict()
     for letter in list(word):
         result[letter] = result.get(letter, 0) + 1
     return result
 
+# returns True if typo is close enough to a word
+## the dissimilarity index is calculated based on letter counts and substring search
 def closeEnough(typo, word, index):
     diff = 0
     typoData, wordData = letterCounts(typo), letterCounts(word)
@@ -34,6 +37,7 @@ def closeEnough(typo, word, index):
             diff += 1
     return diff / (2 * len(word)) < index
 
+# returns the closest word to the typo
 def findMatches(typo, words, index):
     for word in words:
         if closeEnough(typo, word, index):
@@ -42,6 +46,7 @@ def findMatches(typo, words, index):
 
 # command recognition
 
+# returns callsign that matches the callsign in the command
 def findCallsign(cmd, aircrafts):
     word = cmd.split(' ')
     for aircraft in aircrafts:
@@ -59,6 +64,7 @@ def findCallsign(cmd, aircrafts):
         if word[0] == no:
             return aircraft.callsign
             
+# finds and returns numerical altitude value
 def findAltitude(words):
     numbers = []
     for word in words:
@@ -68,6 +74,7 @@ def findAltitude(words):
         if int(number) % 50 == 0:
             return number
 
+# returns True if an altitude command is detected in command string
 def findAltCommand(cmd):
     words = cmd.split(" ")
     keywords = ["climb", "descend"]
@@ -77,6 +84,7 @@ def findAltCommand(cmd):
             return findAltitude(altComm)
     return None
 
+# finds and returns numerical heading value
 def findHeading(words):
     numbers = []
     for word in words:
@@ -86,6 +94,7 @@ def findHeading(words):
         if 0 <= int(number) <= 360:
             return number
 
+# returns True if an heading command is detected in command string
 def findHdgCommand(cmd):
     words = cmd.split(" ")
     keywords = ["heading", "turn", "right", "left"]
@@ -95,6 +104,7 @@ def findHdgCommand(cmd):
             return findHeading(hdgComm)
     return None
 
+# finds and returns numerical speed value
 def findSpeed(words):
     numbers = []
     for word in words:
@@ -102,9 +112,10 @@ def findSpeed(words):
             numbers.append(word)
     for number in numbers:
         # speed is under 250 kts below 10000 ft
-        if 130 <= int(number) < 250 and int(number) % 5 == 0:
+        if 130 <= int(number) <= 250 and int(number) % 5 == 0:
             return number
 
+# finds and returns alphabetical waypoint object based on name
 def findWaypoint(words, airport):
     names = []
     for word in words:
@@ -116,6 +127,7 @@ def findWaypoint(words, airport):
                 return waypoint
     return None
 
+# returns True if an direct command is detected in command string
 def findDirectCommand(cmd, airport):
     words = cmd.split(" ")
     keywords = ["direct", "waypoint"]
@@ -125,15 +137,17 @@ def findDirectCommand(cmd, airport):
             return findWaypoint(wptComm, airport)
     return None
 
+# returns True if an speed command is detected in command string
 def findSpdCommand(cmd):
     words = cmd.split(" ")
-    keywords = ["accelerate", "decelerate", "speed", "decel"]
+    keywords = ["accelerate", "decelerate", "speed", "decel", "accel"]
     for word in words:
         if findMatches(word, keywords, 0.4) != None:
             spdComm = words[words.index(word):]
             return findSpeed(spdComm)
     return None
 
+# returns True if an altitude command is detected in command string
 def findTakeoffClearance(cmd):
     words = cmd.split(" ")
     keywords = ["cleared", "takeoff"]
@@ -142,9 +156,9 @@ def findTakeoffClearance(cmd):
             return False
     return True
 
+# returns True if debug command is detected in command
 def debugCommand(cmd):
-    words = cmd.split(" ")
-    return ("debug" in words)
+    return ("debug" in cmd.split(" "))
 
 # keyword based command recognition
 def divideCommand(cmd, flights, airport):
@@ -156,10 +170,8 @@ def divideCommand(cmd, flights, airport):
     spd = findSpdCommand(cmd)
     return (callsign, alt, wpt, hdg, spd, clr)
 
+# executes instant change debug commands
 def debugExecuteCommand(flights, airport, cmd):
-    if cmd == "crash":
-        random.choice(flights).crash = True
-        return
     fuel = None
     if "fuel" in cmd:
         fuel = 3000
@@ -174,7 +186,11 @@ def debugExecuteCommand(flights, airport, cmd):
         if hdg != None: flight.hdg = int(hdg)
         if spd != None: flight.spd = int(spd)
 
+# executes real time change commands
 def executeCommand(flights, airport, cmd):
+    if cmd == "crash":
+        random.choice(flights).crash = True
+        return
     (callsign, alt, wpt, hdg, spd, clr) = divideCommand(cmd, flights, airport)
     flight = None
     for fl in flights: